[Refactor] : chat. controller 패키지 내의 dro 패키지 제거
[Refactor] : chat. controller 패키지 내의 dro 패키지 제거
